pyproject.toml
[tool.poetry]
name = "ml-based-anomaly-detector"
version = "0.1.0"
description = ""
authors = ["Thomas Selin"]
readme = "README.md"
packages = [
  {include = "ml_monitoring_service", from = "src"},
]
license = "MIT"

[tool.poetry.dependencies]
python-dotenv = "^1.1.1"
python = ">=3.12,<3.14"
numpy = "^1.26.0"
pandas = "^2.2.3"
requests = "^2.32.3"
urllib3 = "^2.2.3"
psutil = "^6.1.1"
torch-geometric = "^2.6.1"
networkx = "^3.4.2"
matplotlib = "^3.9.2"
torchviz = "^0.0.2"
graphviz = "^0.20.3"
apscheduler = "^3.11.0"
flask = "^3.1.0"
pytz = "^2024.2"
torchvision = "^0.20.1"
pyyaml = "^6.0.2"
torchview = "^0.2.6"
mlflow = ">=2.11.0,<3.0.0"
boto3 = "^1.37.28"
pytest-env = "^1.1.5"
waitress = "^3.0.2"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"
pytest-flask = "^1.3.0"
pytest-env = "^1.1.5"
pre-commit = "^4.0.1"
ruff = "^0.9.2"
mypy = "^1.14.1"
pyright = "^1.1.407"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "--cov=src --cov-branch --cov-report=term-missing tests"
env = [
    "MLFLOW_EXPERIMENT_NAME=ML Based Anomaly Detector: Unit/pytest runs"
]

[tool.ruff]
target-version = "py312"
line-length = 88

extend-exclude = [
  "mlruns",
  "output",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501", "E402"]

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = false
disallow_untyped_defs = false
no_implicit_optional = false
ignore_missing_imports = true
exclude = "^(mlruns|output)/"
